requirements.txt
ray
dm-reverb
tensorflow
numpy
matplotlib
gym